pkg/executor: isolate recover heavy tests in dedicated suite
