+++ pkg/executor/test/recovertest/BUILD.bazel
@@ -1,0 +1,29 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_test")
+
+go_test(
+    name = "recovertest_test",
+    timeout = "short",
+    srcs = [
+        "main_test.go",
+        "recover_test.go",
+    ],
+    flaky = True,
+    shard_count = 2,
+    deps = [
+        "//pkg/config",
+        "//pkg/ddl/util",
+        "//pkg/kv",
+        "//pkg/meta/autoid",
+        "//pkg/store/mockstore",
+        "//pkg/testkit",
+        "//pkg/testkit/testfailpoint",
+        "//pkg/util",
+        "//pkg/util/gcutil",
+        "@com_github_stretchr_testify//require",
+        "@com_github_tikv_client_go_v2//oracle",
+        "@com_github_tikv_client_go_v2//tikv",
+        "@com_github_tikv_client_go_v2//util",
+        "@io_opencensus_go//stats/view",
+        "@org_uber_go_goleak//:goleak",
+    ],
+)
